Added --index switch to index all subdirectories of the current dir.
Known bug: Directories lower in the directory tree will receive a higher rating, probably due to the way the _z --add command works.

Caution is advised when using this, but I think this or something similar would probably be a good feature for z, as long as you know what you are getting yourself into.