Edit the Installation section so that it reads better
Having the "Optionally:" section straight after "Installation:" made it look as if the Installation section had been accidentally deleted. Also the word "Optionally" suggests that you had previously been given something for this to be an option for. e.g. "optionally you can also do this" or "optionally you can do this instead".