Alternate approach for appending to PROMPT_COMMAND
Our basic problem with appending to the end of $PROMPT_COMMAND was that
we couldn't put 2 semicolons in a row.  We can, however, put 2 newlines
in a row, which seems to be treated identically.  As far as I can tell,
this approach works whether PROMPT_COMMAND is unset, set to an empty
string, or set to a non-empty string with or without a trailing
semicolon.