Partially fix '-x' argument
Previously, on OS X, the command `sed -i -e` was used incorrectly and made a backup file named ".ze". Using `sed -i ''` prevents this backup file from being created (on OS X).

Previously, after running `z -x`, the datafile would be printed as if `z` failed to find results for a query. But, because `z -x` would never be used with args (i.e., a query), it is best to `return` like the `-h` command.

However, the `z -x` still does not work because of `precmd`, which adds the current directory to the datafile after every `z` run regardless of the arguments. `z -x` correctly modifies the datafile, but after, `z` add the current directory back.